Globbing: Switch to newer globbing solution
Globbing2 is replacing Globbing in-place, with minor
changes to the globbing syntax.
